Add first trade, last trade timestamp and total trades fields

+++ analytics/src/main/java/com/crypto/analytics/mapper/ZonedDateTimeMapper.java
@@ -15,7 +15,7 @@ public abstract class ZonedDateTimeMapper {
     public static final String FROM_UNIX_TIMESTAMP = "fromUnixTimestamp";
 
     @Named(FROM_UNIX_TIMESTAMP)
-    public ZonedDateTime fromUnixTimestamp(Integer timestamp) {
+    public ZonedDateTime fromUnixTimestamp(Long timestamp) {
         if (isNull(timestamp)) {
             return null;
         }
